server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.101
redis==5.0.1